Move VGSTextRange to a separate file | add more logs
